# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste API",
        "description": "Teste LIVRO",
        "version": "1.0.0"
    },
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Lista todos os autores",
                "description": "Retorna uma coleção paginada de autores",
                "operationId": "0c71439e213e1b284fc37995d74bca2b",
                "responses": {
                    "200": {
                        "description": "Lista de autores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "cod_au": {
                                                "description": "Código do Autor",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "description": "Nome do Autor",
                                                "type": "string",
                                                "example": "Main Branch"
                                            },
                                            "created_at": {
                                                "description": "Data e hora de criação do registro",
                                                "type": "string",
                                                "example": "01/01/2025 12:00:00"
                                            },
                                            "updated_at": {
                                                "description": "Data e hora de atualização do registro",
                                                "type": "string",
                                                "example": "01/02/2025 15:30:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Cria um novo autor",
                "description": "Cria um autor a partir dos dados fornecidos",
                "operationId": "291d8539d3b5490d46fe9f2ebd84c64c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "José da Silva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_au": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "José da Silva"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou"
                    }
                }
            }
        },
        "/api/authors/{cod_au}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Retorna um autor pelo Codido do Autor",
                "operationId": "ed472fc79887fac0172d68ad62a2d430",
                "parameters": [
                    {
                        "name": "cod_au",
                        "in": "path",
                        "description": "Código do autor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do autor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cod_au": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "José da Silva"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-06-02T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Atualiza um autor existente",
                "operationId": "d086b836b8ca567a225c65a523fc5e0a",
                "parameters": [
                    {
                        "name": "cod_au",
                        "in": "path",
                        "description": "Código do autor a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "José da Silva"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Autor atualizado com sucesso, sem conteúdo retornado"
                    },
                    "422": {
                        "description": "Falha na validação"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Deleta um autor pelo código do autor",
                "operationId": "6eec67d84716b4c4b5dfefac47c3fd51",
                "parameters": [
                    {
                        "name": "cod_au",
                        "in": "path",
                        "description": "Código do autor a ser deletado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor deletado com sucesso"
                    },
                    "404": {
                        "description": "Autor não encontrado"
                    }
                }
            }
        },
...
    },
...
}